remove stale file
