+++ internal/engine/local/controller.go
@@ -3,6 +3,7 @@ package local
 import (
 	"context"
 	"fmt"
+	"time"
 
 	"github.com/windmilleng/tilt/internal/store"
 	"github.com/windmilleng/tilt/pkg/logger"
@@ -169,3 +170,19 @@ func (w LocalServeLogActionWriter) Write(level logger.Level, p []byte) error {
 func SpanIDForServeLog(sequenceNum int) logstore.SpanID {
 	return logstore.SpanID(fmt.Sprintf("localserve:%d", sequenceNum))
 }
+
+// ServeSpec describes what Runner should be running
+type ServeSpec struct {
+	ManifestName model.ManifestName
+	ServeCmd     model.Cmd
+	TriggerTime  time.Time // TriggerTime is how Runner knows to restart; if it's newer than the TriggerTime of the currently running command, then Runner should restart it
+}
+
+type Status int
+
+const (
+	Unknown Status = iota
+	Running
+	Done
+	Error
+)
